requirements.txt
Django>=5.1.0,<5.2.0
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.6,<2.10.0
Pillow>=9.5.0,<10.0.0
flake8>=7.1.0,<7.120
